autocon-server/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
Pillow>=10.0,<12.0